backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
networkx
pandas
requests